Move TextWithHTML to components/ui
