--- airbyte-webapp/src/components/TextWithHTML/index.tsx
@@ -1,4 +1,0 @@
-import TextWithHTML from "./TextWithHTML";
-
-export default TextWithHTML;
-export { TextWithHTML };
