+++ airbyte-webapp/src/components/ui/TextWithHTML/index.stories.tsx
@@ -1,0 +1,13 @@
+import { ComponentStory, ComponentMeta } from "@storybook/react";
+
+import { TextWithHTML } from "./TextWithHTML";
+
+export default {
+  title: "UI/TextWithHTML",
+  component: TextWithHTML,
+  argTypes: {},
+} as ComponentMeta<typeof TextWithHTML>;
+
+const Template: ComponentStory<typeof TextWithHTML> = (args) => <TextWithHTML {...args} />;
+
+export const Primary = Template.bind({});
